requirements.txt
actionlib==1.12.0
angles==1.9.11
attrs==19.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.4
bondpy==1.8.3
camera-calibration==1.13.0
camera-calibration-parsers==1.11.13
catkin==0.7.19
catkin-pkg==0.4.16
certifi==2020.4.5.2
chardet==3.0.4
cloudpickle==1.3.0
configparser==4.0.2
contextlib2==0.6.0.post1
controller-manager==0.15.1
controller-manager-msgs==0.15.1
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.16
dynamic-reconfigure==1.6.0
entrypoints==0.3
enum34==1.1.10
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
gazebo-plugins==2.8.4
gazebo-ros==2.8.4
gencpp==0.6.2
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.12
gennodejs==2.0.1
genpy==0.6.9
idna==2.9
image-geometry==1.13.0
importlib-metadata==1.6.0
interactive-markers==1.11.4
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.3
kdl-parser-py==1.13.1
kiwisolver==1.1.0
laser-geometry==1.6.4
MarkupSafe==1.1.1
matplotlib==2.2.5
message-filters==1.14.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.8
numpy==1.16.6
opencv-python==4.2.0.32
pandocfilters==1.4.2
pathlib==1.0.1
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
prometheus-client==0.7.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
pycocotools==2.0
Pygments==2.5.2
pyparsing==2.4.7
pypcd==0.1.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-lzf==0.2.4
python-qt-binding==0.3.6
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.3.1
pyzmq==19.0.0
qt-dotgraph==0.3.16
qt-gui==0.3.16
qt-gui-cpp==0.3.16
qt-gui-py-common==0.3.16
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
resource-retriever==1.12.5
rosbag==1.14.3
rosboost-cfg==1.14.7
rosclean==1.14.7
roscreate==1.14.7
rosgraph==1.14.3
roslaunch==1.14.3
roslib==1.14.7
roslint==0.11.2
roslz4==1.14.3
rosmake==1.14.7
rosmaster==1.14.3
rosmsg==1.14.3
rosnode==1.14.3
rosparam==1.14.3
rospy==1.14.3
rosservice==1.14.3
rostest==1.14.3
rostopic==1.14.3
rosunit==1.14.7
roswtf==1.14.3
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.10
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.9
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.13.6
scandir==1.10.0
scikit-image==0.14.5
scipy==1.2.3
Send2Trash==1.5.0
sensor-msgs==1.12.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
subprocess32==3.5.4
terminado==0.8.3
testpath==0.4.4
tf==1.12.0
tf-conversions==1.12.0
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.3
torch==1.4.0
torchvision==0.5.0
tornado==5.1.1
traitlets==4.3.3
urdfdom-py==0.4.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
xacro==1.13.4
zipp==1.2.0
